Rename drop endpoint to empty and nest queue filters
Rename prompt/drop to prompt/empty and update router imports and handler names for clarity. Files changed: app/api/prompt/drop.py -> app/api/prompt/empty.py (endpoint path /prompt/empty, handler empty_prompt_table, removed unused os import), app/api/prompt/__init__.py and app/api/routes.py (update imports and router include names). Also adjust queue GET response shape (app/api/queue/routes/get.py) to group collection/group values and collections/groups under a new "filters" object with keys collectionFilter and groupFilter. Clients should be updated to use the new endpoint and the revised queue response structure.
