+++ app/api/prompt/__init__.py
@@ -1,6 +1,5 @@
 """Prompt Routes"""
 
-
 from .prompt import router as prompt_router
 from .linkedin import router as linkedin_router
-from .drop import router as drop_router
+from .empty import router as empty_router
